refact

+++ sources/Device/src/Menu/Pages/PageService/PageDebug/PageDebug_ADC.cpp
@@ -132,8 +132,8 @@ void PageDebug::PageADC::PageStretch::OnChanged_Mode(bool)
     }
     else
     {
-        stretchA = set.dbg.nrst.stretchADC[Chan::A];
-        stretchB = set.dbg.nrst.stretchADC[Chan::B];
+        stretchA = static_cast<int16>(set.dbg.nrst.stretchADC[Chan::A]);
+        stretchB = static_cast<int16>(set.dbg.nrst.stretchADC[Chan::B]);
     }
 }
